Desktop Bug fixes
I added tests for the lodging DAL and the trip, Transportation, Waypoint, User, and Lodging model classes. Added the Code Coverage xaml file to the test code coverage folder inside the Sprint 2 folder. Modified the Overlapping Waypoint and Transportation error messages to show the overlapping waypoints/Transportations. Named the back button in the TransportationInfo code behind file so that the system will no longer crash when a transport is removed.